+++ pkg/cstorpool/v1alpha2/kubernetes.go
@@ -1,84 +1,149 @@
-// Copyright © 2018-2019 The OpenEBS Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+Copyright 2019 The OpenEBS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 
 package v1alpha2
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"strings"
 
-	kclient "github.com/openebs/maya/pkg/kubernetes/client/v1alpha1"
+	"k8s.io/apimachinery/pkg/types"
 
 	apis "github.com/openebs/maya/pkg/apis/openebs.io/v1alpha1"
 	clientset "github.com/openebs/maya/pkg/client/generated/clientset/versioned"
+	errors "github.com/openebs/maya/pkg/errors/v1alpha1"
+	client "github.com/openebs/maya/pkg/kubernetes/client/v1alpha1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// getClientsetFn is a typed function that
-// abstracts fetching of internal clientset
-type getClientsetFn func() (clientset *clientset.Clientset, err error)
+// getClientsetFunc is a typed function that
+// abstracts fetching internal clientset
+type getClientsetFunc func() (cs *clientset.Clientset, err error)
 
-// listFn is a typed function that abstracts
-// listing of cstor pool
-type listFn func(cli *clientset.Clientset, opts metav1.ListOptions) (*apis.CStorPoolList, error)
+// listFunc is a typed function that abstracts
+// listing CStorPool instances
+type listFunc func(cs *clientset.Clientset, opts metav1.ListOptions) (*apis.CStorPoolList, error)
 
-// kubeclient enables kubernetes API operations
-// on cstor storage pool instance
-type kubeclient struct {
-	// clientset refers to cstor storage pool's
+// getFunc is a typed function that abstracts
+// getting CStorPool instances
+type getFunc func(cs *clientset.Clientset, name string, opts metav1.GetOptions) (*apis.CStorPool, error)
+
+// createFunc is a typed function that abstracts
+// creating CStorPool instances
+type createFunc func(cs *clientset.Clientset, storegePoolObj *apis.CStorPool) (*apis.CStorPool, error)
+
+// patchFunc is a typed function that abstracts
+// patching CStorPool instances
+type patchFunc func(cs *clientset.Clientset, name string, pt types.PatchType, patchObj []byte) (*apis.CStorPool, error)
+
+// delFn is a typed function that abstracts
+// delete of StoragePool instances
+type delFn func(cs *clientset.Clientset, name string,
+	opts *metav1.DeleteOptions) error
+
+// Kubeclient enables kubernetes API operations
+// on CStorPool instance
+type Kubeclient struct {
+	// clientset refers to CStorPool's
 	// clientset that will be responsible to
 	// make kubernetes API calls
 	clientset *clientset.Clientset
-
 	// functions useful during mocking
-	getClientset getClientsetFn
-	list         listFn
+	getClientset getClientsetFunc
+	list         listFunc
+	get          getFunc
+	create       createFunc
+	patch        patchFunc
+	del          delFn
 }
 
-// kubeclientBuildOption defines the abstraction
-// to build a kubeclient instance
-type kubeclientBuildOption func(*kubeclient)
+// KubeclientBuildOption defines the abstraction
+// to build a Kubeclient instance
+type KubeclientBuildOption func(*Kubeclient)
 
 // withDefaults sets the default options
 // of kubeclient instance
-func (k *kubeclient) withDefaults() {
+func (k *Kubeclient) withDefaults() {
 	if k.getClientset == nil {
-		k.getClientset = func() (clients *clientset.Clientset, err error) {
-			config, err := kclient.New().Config()
+		k.getClientset = func() (cs *clientset.Clientset, err error) {
+			config, err := client.GetConfig(client.New())
 			if err != nil {
 				return nil, err
 			}
 			return clientset.NewForConfig(config)
 		}
 	}
+
 	if k.list == nil {
-		k.list = func(cli *clientset.Clientset, opts metav1.ListOptions) (*apis.CStorPoolList, error) {
-			return cli.OpenebsV1alpha1().CStorPools().List(opts)
+		k.list = func(cs *clientset.Clientset,
+			opts metav1.ListOptions) (*apis.CStorPoolList, error) {
+			return cs.OpenebsV1alpha1().
+				CStorPools().
+				List(opts)
+		}
+	}
+
+	if k.get == nil {
+		k.get = func(cs *clientset.Clientset,
+			name string, opts metav1.GetOptions) (*apis.CStorPool, error) {
+			return cs.OpenebsV1alpha1().
+				CStorPools().
+				Get(name, opts)
+		}
+	}
+
+	if k.create == nil {
+		k.create = func(cs *clientset.Clientset,
+			obj *apis.CStorPool) (*apis.CStorPool, error) {
+			return cs.OpenebsV1alpha1().
+				CStorPools().
+				Create(obj)
+		}
+	}
+
+	if k.patch == nil {
+		k.patch = func(cs *clientset.Clientset, name string,
+			pt types.PatchType, patchObj []byte) (*apis.CStorPool, error) {
+			return cs.OpenebsV1alpha1().
+				CStorPools().
+				Patch(name, pt, patchObj)
+		}
+	}
+
+	if k.del == nil {
+		k.del = func(cs *clientset.Clientset, name string,
+			opts *metav1.DeleteOptions) error {
+			return cs.OpenebsV1alpha1().
+				CStorPools().
+				Delete(name, opts)
 		}
 	}
 }
 
-// WithKubeClient sets the kubernetes client against
+// WithClientset sets the kubernetes clientset against
 // the kubeclient instance
-func WithKubeClient(c *clientset.Clientset) kubeclientBuildOption {
-	return func(k *kubeclient) {
+func WithClientset(c *clientset.Clientset) KubeclientBuildOption {
+	return func(k *Kubeclient) {
 		k.clientset = c
 	}
 }
 
-// KubeClient returns a new instance of kubeclient meant for
-// cstor volume replica operations
-func KubeClient(opts ...kubeclientBuildOption) *kubeclient {
-	k := &kubeclient{}
+// NewKubeClient returns a new instance of kubeclient meant for
+// CStorPool related operations
+func NewKubeClient(opts ...KubeclientBuildOption) *Kubeclient {
+	k := &Kubeclient{}
 	for _, o := range opts {
 		o(k)
 	}
@@ -88,24 +153,71 @@ func KubeClient(opts ...kubeclientBuildOption) *kubeclient {
 
 // getClientOrCached returns either a new instance
 // of kubernetes client or its cached copy
-func (k *kubeclient) getClientOrCached() (*clientset.Clientset, error) {
+func (k *Kubeclient) getClientOrCached() (*clientset.Clientset, error) {
 	if k.clientset != nil {
 		return k.clientset, nil
 	}
 	c, err := k.getClientset()
 	if err != nil {
-		return nil, err
+		return nil,
+			errors.WithStack(errors.Wrapf(err, "failed to get clientset:"))
 	}
 	k.clientset = c
 	return k.clientset, nil
 }
 
-// List returns a list of cstor pool
+// List returns a list of CStorPool
 // instances present in kubernetes cluster
-func (k *kubeclient) List(opts metav1.ListOptions) (*apis.CStorPoolList, error) {
-	cli, err := k.getClientOrCached()
+func (k *Kubeclient) List(opts metav1.ListOptions) (*apis.CStorPoolList, error) {
+	cs, err := k.getClientOrCached()
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "failed to list CStorPool:")
 	}
-	return k.list(cli, opts)
+	return k.list(cs, opts)
+}
+
+// Get returns an CStorPool instance from kubernetes cluster
+func (k *Kubeclient) Get(name string, opts metav1.GetOptions) (*apis.CStorPool, error) {
+	if strings.TrimSpace(name) == "" {
+		return nil, errors.New("failed to get CStorPool: missing name")
+	}
+	cs, err := k.getClientOrCached()
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to get CStorPool: {%s}", name)
+	}
+	return k.get(cs, name, opts)
+}
+
+// Create creates an CStorPool instance in kubernetes cluster
+func (k *Kubeclient) Create(obj *apis.CStorPool) (*apis.CStorPool, error) {
+	cs, err := k.getClientOrCached()
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create CStorPool:")
+	}
+	return k.create(cs, obj)
+}
+
+// Patch returns the patched CStorPool instance
+func (k *Kubeclient) Patch(name string, pt types.PatchType,
+	patchObj []byte) (*apis.CStorPool, error) {
+	if strings.TrimSpace(name) == "" {
+		return nil, errors.New("failed to patch CStorPool: missing name")
+	}
+	cs, err := k.getClientOrCached()
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to patch CStorPool:")
+	}
+	return k.patch(cs, name, pt, patchObj)
+}
+
+// Delete deletes CStorPool instance
+func (k *Kubeclient) Delete(name string, opts *metav1.DeleteOptions) error {
+	if strings.TrimSpace(name) == "" {
+		return errors.New("failed to delete CStorPool: missing name")
+	}
+	cs, err := k.getClientOrCached()
+	if err != nil {
+		return errors.Wrap(err, "failed to delete CStorPool:")
+	}
+	return k.del(cs, name, opts)
 }
